fix: make travel_time func for api configuratble, #ROU-13
