+++ routor/api/main.py
@@ -50,6 +50,7 @@ def read_weights() -> List[str]:
 def read_route(
     data: models.RouteRequest,
     engine: engine.Engine = Depends(get_engine),  # noqa: B008
+    settings: config.Settings = Depends(get_settings),  # noqa: B008
 ) -> engine_models.Route:
     """
     Calculate a route from A to B.
@@ -57,6 +58,6 @@ def read_route(
     weight_func = weights.get_function(data.weight)
 
     route = engine.route(
-        data.origin, data.destination, weight_func, weights.travel_time
+        data.origin, data.destination, weight_func, settings.get_travel_time_func()
     )
     return route
